remove utils/pagerduty references
