+++ pkg/provider/pagerduty/pagerduty.go
@@ -3,6 +3,9 @@ package pagerduty
 import (
 	"context"
 	"fmt"
+	"sort"
+	"strings"
+	"time"
 
 	pd "github.com/PagerDuty/go-pagerduty"
 )
@@ -12,6 +15,12 @@ const (
 	PagerDutyOauthTokenConfigKey = "pd_oauth_token"
 	PagerDutyTeamIDsKey          = "team_ids"
 )
+
+type IncidentOccurrenceTracker struct {
+	IncidentName   string
+	Count          int
+	LastOccurrence string
+}
 
 type pdClientInterface interface {
 	ListIncidentsWithContext(context.Context, pd.ListIncidentsOptions) (*pd.ListIncidentsResponse, error)
@@ -117,4 +126,87 @@ func (c *client) GetFiringAlertsForCluster(pdServiceIDs []string) (map[string][]
 		}
 	}
 	return incidents, nil
+}
+
+func (c *client) GetHistoricalAlertsForCluster(pdServiceIDs []string) (map[string][]*IncidentOccurrenceTracker, error) {
+
+	var currentOffset uint
+	var limit uint = 100
+	var incidents []pd.Incident
+	var ctx = context.TODO()
+	incidentMap := map[string][]*IncidentOccurrenceTracker{}
+
+	for _, pdServiceID := range pdServiceIDs {
+		for currentOffset = 0; true; currentOffset += limit {
+			liResponse, err := c.pdclient.ListIncidentsWithContext(
+				ctx,
+				pd.ListIncidentsOptions{
+					ServiceIDs: []string{pdServiceID},
+					Statuses:   []string{"resolved", "triggered", "acknowledged"},
+					Offset:     currentOffset,
+					Limit:      limit,
+					SortBy:     "created_at:desc",
+				},
+			)
+
+			if err != nil {
+				return nil, err
+			}
+
+			if len(liResponse.Incidents) == 0 {
+				break
+			}
+
+			incidents = append(incidents, liResponse.Incidents...)
+		}
+
+		incidentCounter := make(map[string]*IncidentOccurrenceTracker)
+
+		for _, incident := range incidents {
+			title := strings.Split(incident.Title, " ")[0]
+			if _, found := incidentCounter[title]; found {
+				incidentCounter[title].Count++
+
+				// Compare current incident timestamp vs our previous 'latest occurrence', and save the most recent.
+				currentLastOccurrence, err := time.Parse(time.RFC3339, incidentCounter[title].LastOccurrence)
+				if err != nil {
+					fmt.Printf("Failed to parse time %q\n", err)
+					return nil, err
+				}
+
+				incidentCreatedAt, err := time.Parse(time.RFC3339, incident.CreatedAt)
+				if err != nil {
+					fmt.Printf("Failed to parse time %q\n", err)
+					return nil, err
+				}
+
+				// We want to see when the latest occurrence was
+				if incidentCreatedAt.After(currentLastOccurrence) {
+					incidentCounter[title].LastOccurrence = incident.CreatedAt
+				}
+
+			} else {
+				// First time encountering this incident type
+				incidentCounter[title] = &IncidentOccurrenceTracker{
+					IncidentName:   title,
+					Count:          1,
+					LastOccurrence: incident.CreatedAt,
+				}
+			}
+		}
+
+		var incidentSlice []*IncidentOccurrenceTracker = make([]*IncidentOccurrenceTracker, 0, len(incidentCounter))
+		for _, val := range incidentCounter {
+			incidentSlice = append(incidentSlice, val)
+		}
+
+		sort.Slice(incidentSlice, func(i int, j int) bool {
+			return incidentSlice[i].Count < incidentSlice[j].Count
+		})
+		incidentMap[pdServiceID] = append(incidentMap[pdServiceID], incidentSlice...)
+
+	}
+
+	return incidentMap, nil
+
 }
